fix <unk> and newline token issues
