+++ .gitignore
@@ -3,9 +3,11 @@ data/
 *.pt
 *.pth
 tokens
-AOCHIDES/
 __MACOSX/
 __pycache__/
 *.zip
 babylm_data/
 sentences
+.vscode/
+AOCHILDES/
+babylm/
