--- babylm_tokens.py
@@ -1,160 +1,0 @@
-from nltk import word_tokenize
-import nltk
-from nltk.tokenize import SyllableTokenizer
-from hmm_syllables import *
-from inference import generate_babytalk_sentences
-from trainer import *
-from collections import Counter, OrderedDict
-from sklearn.model_selection import train_test_split
-import torch.utils.data
-import stanza
-import pickle
-from tqdm import tqdm  # for displaying progress bar
-import numpy as np
-import torch
-from aochildes.dataset import AOChildesDataSet
-from pathlib import Path
-import os
-
-nltk.download('punkt')
-
-
-def encode(s, vocab):
-    """
-    Convert a string into a list of integers
-    """
-    x = sum([SSP.tokenize(word)+[" "]
-            for word in word_tokenize(s)], []) + ["\n"]
-    return [vocab[xx] for xx in x]
-
-
-def decode(x, vocab):
-    """
-    Convert list of ints to string
-    """
-    s = " ".join([vocab.lookup_token(xx) for xx in x])
-    return s
-
-
-# transcripts = AOChildesDataSet().load_transcripts()
-
-
-class TextDataset(torch.utils.data.Dataset):
-    def __init__(self, lines, vocab):
-        self.lines = lines  # list of strings
-        # function for generating a minibatch from strings
-        collate = Collate(vocab=vocab)
-        self.loader = torch.utils.data.DataLoader(
-            self, batch_size=32, shuffle=True, collate_fn=collate)
-
-    def __len__(self):
-        return len(self.lines)
-
-    def __getitem__(self, idx):
-        line = self.lines[idx]
-        return line
-
-
-class Collate:
-    def __init__(self, vocab):
-        self.vocab = vocab
-
-    def __call__(self, batch):
-        """
-        Returns a minibatch of strings, padded to have the same length.
-        """
-        x = []
-        batch_size = len(batch)
-        for index in range(batch_size):
-            x_ = batch[index]
-            # convert letters to integers
-            x.append(encode(x_, self.vocab))
-        # pad all sequences with 0 to have same length
-        x_lengths = [len(x_) for x_ in x]
-        T = max(x_lengths)
-        for index in range(batch_size):
-            x[index] += [0] * (T - len(x[index]))
-            x[index] = torch.tensor(x[index])
-
-        # stack into single tensor
-        x = torch.stack(x)
-        x_lengths = torch.tensor(x_lengths)
-        return (x, x_lengths)
-
-
-""" for i, text in enumerate(tqdm(transcripts)):
-  for sent in nlp(text).sentences:
-    for token in sent.tokens:
-      tokens.append(token.text) """
-
-"""
-with open("tokens", "wb") as fp:   #Pickling
-   pickle.dump(tokens, fp) """
-
-
-def train_hmm_syllables():
-    transcripts = AOChildesDataSet().load_transcripts()
-    SSP = SyllableTokenizer()
-    vocab = build_or_load_vocab(transcripts)
-    vocab.insert_token(" ", len(vocab))
-    vocab.set_default_index(vocab['<unk>'])
-
-    unk_token = '<unk>'
-    nlp = stanza.Pipeline(lang='en', processors='tokenize',
-                          tokenize_no_ssplit=True)
-
-    sentences = []
-    for text in tqdm(transcripts):
-        sentences += split_into_sentences(text)
-
-    with open("sentences", "wb") as fp:  # Pickling
-        pickle.dump(sentences, fp)
-    print("**********************")
-    with open("sentences", "rb") as fp:   # Unpickling
-        sentences = pickle.load(fp)
-    print(sentences[:10])
-    train_lines, valid_lines = train_test_split(
-        sentences, test_size=0.1, random_state=42)
-    train_dataset = TextDataset(train_lines, vocab)
-
-    model = HMM_syllable(M=len(vocab), N=6)
-    optimizer = torch.optim.Adam(
-        model.parameters(), lr=0.01, weight_decay=0.00001)
-    train_loss = 0
-    num_samples = 0
-    model.train()
-    print_interval = 50
-
-    for idx, batch in enumerate(tqdm(train_dataset.loader)):
-        x, T = batch
-        batch_size = len(x)
-        num_samples += batch_size
-        log_probs = model(x, T)
-        loss = -log_probs.mean()
-        optimizer.zero_grad()
-        loss.backward()
-        optimizer.step()
-        train_loss += loss.cpu().data.numpy().item() * batch_size
-        if idx % print_interval == 0:
-            print("loss:", loss.item())
-            for _ in range(5):
-                sampled_x, sampled_z = model.sample(
-                    stop_token_index=vocab["\n"])
-                print(sampled_z)
-                print(decode(sampled_x, vocab))
-
-    torch.save(model.state_dict(), "model_tokens.pt")
-    # model = HMM_syllable(M=len(vocab), N=6)
-    # model.load_state_dict(torch.load("model_tokens.pt"))
-    print("saved")
-
-
-if __name__ == "__main__":
-    import aochildes
-    aochildes.configs.Dirs.transcripts = Path("data/aochildes/")
-    # train_hmm_syllables()
-    # FIXME (Leon): 1. why is <unk> in output?
-    # 2. why are there so many newlines?
-    # 3. Remove newlines from vocab
-    # 4. make sure the ouput of the below call is exactly 10 sentences
-    generate_babytalk_sentences(10, "babytalk_corpus.txt")
