+++ requirements.txt
@@ -2,9 +2,9 @@ fitter
 torch
 numpy
 pyprind
-sklearn
 nltk
 torchtext
 stanza
 git+https://github.com/UIUCLearningLanguageLab/AOCHILDES.git
 gdown
+scikit-learn
